+++ src/coordio/__init__.py
@@ -34,6 +34,7 @@ except OSError:
     sofa = None
 
 
+from .defaults import calibration
 from .exceptions import *
 from .iers import IERS
 from .site import Site