+++ CHANGELOG.rst
@@ -10,8 +10,9 @@ Change Log
 
 * Use measured alpha and beta offsets when transforming from tangent to positioner.
 * Replace error in ``iauPmsafe`` with warning.
-* Add ``RoughTransform`` and ``ZhaoBurgeTransform``.
-* Undo changes to ``wokToTangentArr``. Reverted to only supporting one holeID per array.
+* Add ``RoughTransform`` and ``ZhaoBurgeTransform`` (#11).
+* Undo changes to ``wokToTangentArr``. Reverted to only supporting one holeID per array (#11).
+* Add a ``Calibration`` class to store all active calibrations, allowing for concatenation of different site calibrations (#12).
 
 
 1.1.1 (2021-10-28)